+++ changelog.md
@@ -32,6 +32,7 @@ Codion Change Log
 - EntityTablePanel.table now initialized lazily.
 - EntityTablePanel.delete(), deleteWithConfirmation() and editSelectedEntities() renamed deleteSelected(), deleteSelectedWithConfirmation() and editSelected() respectively.
 - EntityTablePanel.printTable() removed.
+- EntityTablePanel.setReferentialIntegrityErrorHandling() replaced with Value based referentialIntegrityErrorHandling(). 
 ### is.codion.swing.framework.server.monitor
 - ServerMonitor and ServerMonitorPanel, tables migrated from plain JTable to FilteredTable.